# src/CertManagerClient/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CertManager API",
    "version": "v1"
  },
  "paths": {
...
    "/{organization-id}/api/v1/CertificateVersions/{id}": {
      "get": {
        "tags": [
          "CertificateVersion"
        ],
        "operationId": "GetCertificateVersionById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "organization-id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CertificateVersionModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CertificateVersionModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CertificateVersionModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "CertificateVersion"
        ],
        "operationId": "DeleteCertificateVersion",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "organization-id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/{organization-id}/api/v1/CertificateVersions": {
      "get": {
        "tags": [
          "CertificateVersion"
        ],
        "operationId": "GetCertificateVersions",
        "parameters": [
          {
            "name": "CertificateIds",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "format": "uuid"
              }
            }
          },
          {
            "name": "MinimumUtcExpirationTime",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "MaximumUtcExpirationTime",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "MinimumUtcActivationTime",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "MaximumUtcActivationTime",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "organization-id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CertificateVersionModel"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CertificateVersionModel"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CertificateVersionModel"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CertificateVersionModel": {
        "type": "object",
        "properties": {
          "activationDate": {
            "type": "string",
            "format": "date-time"
          },
          "cn": {
            "type": "string",
            "nullable": true
          },
          "expiryDate": {
            "type": "string",
            "format": "date-time"
          },
          "issuerName": {
            "type": "string",
            "nullable": true
          },
          "thumbprint": {
            "type": "string",
            "nullable": true
          },
          "rawCertificate": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "certificateId": {
            "type": "string",
            "format": "uuid"
          },
          "certificateVersionId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      }
    },
    "securitySchemes": {
      "JWTBearerAuth": {
        "type": "oauth2",
        "description": "Specify the authorization token.",
        "flows": {
          "password": {
            "authorizationUrl": "http://localhost:8080/realms/master/protocol/openid-connect/auth",
            "tokenUrl": "http://localhost:8080/realms/master/protocol/openid-connect/token",
            "scopes": {
              "cert-manager/read": "Read access",
              "cert-manager/write": "write access"
            }
          },
          "clientCredentials": {
            "authorizationUrl": "http://localhost:8080/realms/master/protocol/openid-connect/auth",
            "tokenUrl": "http://localhost:8080/realms/master/protocol/openid-connect/token",
            "scopes": {
              "cert-manager/read": "Read access",
              "cert-manager/write": "write access"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "JWTBearerAuth": [ ]
    }
  ]
}